flakreate: Move bulk of main.rs to lib.rs
